Dermatological_Diagnosis_and_Consultation_System/requirements.txt
Django==4.2.7
Pillow==10.1.0
tensorflow>=2.13.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
